added handler for create transaction

+++ internal/vertical/transaction/infrastructure/repository/gorm/models/models.go
@@ -1,13 +1,13 @@
 package models
 
 type Transaction struct {
-	ID              string `gorm:"primaryKey"`
-	Amount          int    `gorm:"not null"`
-	CommerceID      string `gorm:"not null"`
-	BranchID        string `gorm:"not null"`
-	UserID          string `gorm:"not null"`
-	PointsErned     int    `gorm:"not null"`
-	CashbackEarned  int    `gorm:""`
-	TransactionDate string `gorm:"not null"`
-	CreatedAt       string `gorm:"autoCreateTime"`
+	ID              string  `gorm:"primaryKey"`
+	Amount          float64 `gorm:"not null"`
+	CommerceID      string  `gorm:"not null"`
+	BranchID        string  `gorm:"not null"`
+	UserID          string  `gorm:"not null"`
+	PointsErned     float64 `gorm:"not null"`
+	CashbackEarned  float64 `gorm:""`
+	TransactionDate string  `gorm:"not null"`
+	CreatedAt       string  `gorm:"autoCreateTime"`
 }
